Reorganise panels around 'NodeSuggester'
