+++ package.json
@@ -84,6 +84,7 @@
     "history": "^4.5.1",
     "react": "^15.3.2",
     "react-dom": "^15.3.2",
+    "react-draggable": "^2.2.3",
     "react-hammerjs": "^0.5.0",
     "react-icons": "^2.2.3",
     "react-redux": "^4.4.5",
