+++ src/components/Elements/index.js
@@ -4,3 +4,5 @@ export { default as Node } from './Node';
 export { default as Form } from './Form';
 export { default as Modal } from './Modal';
 export { default as NodeList } from './NodeList';
+export { default as Panels } from './Panels';
+export { default as Panel } from './Panel';
