+++ src/containers/Elements/index.js
@@ -1,3 +1,2 @@
-export { default as Panels } from './Panels';
-export { default as Panel } from './Panel';
+export { default as NodeSuggester } from './NodeSuggester';
 export { default as PromptSwiper } from './PromptSwiper';
