+++ src/containers/Interfaces/NameGenerator.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 
 import { actionCreators as networkActions } from '../../ducks/modules/network';
 
-import { PromptSwiper, Panels } from '../../containers/Elements';
+import { PromptSwiper, NodeSuggester } from '../../containers/Elements';
 import { NodeList, Modal, Form } from '../../components/Elements';
 
 const nodeLabel = function(node) {
@@ -83,12 +83,13 @@ class NameGeneratorInterface extends Component {
     return (
       <div className='interface'>
         <div className='interface__aside'>
-          <Panels panels={ panels } nodes={ nodes } handleSelect={ this.handlePanelSelect } handleDrag={ this.handlePanelDrag }/>
+          <NodeSuggester config={ panels } />
         </div>
         <div className='interface__primary'>
           <PromptSwiper prompts={ prompts } promptIndex={ this.state.promptIndex } handleNext={ this.nextPrompt } handlePrevious={ this.previousPrompt } />
 
           <NodeList nodes={ nodes } label={ nodeLabel } />
+
           <button onClick={this.toggleModal}>
             Add a person
           </button>
